+++ clash-lib/src/Clash/Core/TermLiteral.hs
@@ -6,7 +6,10 @@ Maintainer  :  Christiaan Baaij <christiaan.baaij@gmail.com>
 Tools to convert a 'Term' into its "real" representation
 -}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE QuasiQuotes #-}
 {-# LANGUAGE TemplateHaskell #-}
+
+--{-# OPTIONS_GHC -ddump-splices #-}
 
 module Clash.Core.TermLiteral
   ( TermLiteral
@@ -15,21 +18,25 @@ module Clash.Core.TermLiteral
   , uncheckedTermToData
   ) where
 
-import qualified Data.Text                       as Text
-import           Data.Text                       (Text)
 import           Data.Bifunctor                  (bimap)
 import           Data.Either                     (lefts)
+import           Data.Proxy                      (Proxy(..))
+import           Data.Text                       (Text)
+import qualified Data.Text                       as Text
+import           Data.Typeable                   (Typeable, typeRep)
 import           GHC.Natural
 import           GHC.Stack
 
 import           Clash.Core.Term                 (Term(Literal), collectArgs)
 import           Clash.Core.Literal
 import           Clash.Core.Pretty               (showPpr)
+import qualified Clash.Util.Interpolate          as I
+import qualified Clash.Verification.Internal     as Cv
 
 import           Clash.Core.TermLiteral.TH
 
 -- | Tools to deal with literals encoded as a "Term".
-class TermLiteral a where
+class Typeable a => TermLiteral a where
   -- | Convert 'Term' to the constant it represents. Will return an error if
   -- (one of the subterms) fail to translate.
   termToData
@@ -52,6 +59,11 @@ instance TermLiteral Text where
 
 instance TermLiteral Int where
   termToData (collectArgs -> (_, [Left (Literal (IntLiteral n))])) =
+    Right (fromInteger n)
+  termToData t = Left t
+
+instance TermLiteral Word where
+  termToData (collectArgs -> (_, [Left (Literal (WordLiteral n))])) =
     Right (fromInteger n)
   termToData t = Left t
 
@@ -81,14 +93,35 @@ instance TermLiteral a => TermLiteral (Maybe a) where
 instance TermLiteral Bool where
   termToData = $(deriveTermToData ''Bool)
 
+instance TermLiteral Cv.RenderAs where
+  termToData = $(deriveTermToData ''Cv.RenderAs)
+
+instance TermLiteral a => TermLiteral (Cv.Assertion' a) where
+  termToData = $(deriveTermToData ''Cv.Assertion')
+
+instance TermLiteral a => TermLiteral (Cv.Property' a) where
+  termToData = $(deriveTermToData ''Cv.Property')
+
 -- | Same as 'termToData', but returns printable error message if it couldn't
 -- translate a term.
-termToDataError :: TermLiteral a => Term -> Either String a
+termToDataError :: forall a. TermLiteral a => Term -> Either String a
 termToDataError term = bimap err id (termToData term)
  where
-  err failedTerm =
-    "Failed to translate term to literal. Term that failed to translate:\n\n"
-    ++ showPpr failedTerm ++ "\n\nIn the full term:\n\n" ++ showPpr term
+  typ = show (typeRep (Proxy @a))
+
+  err failedTerm = [I.i|
+    Failed to translate term to literal. Term that failed to translate:
+
+      #{showPpr failedTerm}
+
+    In the full term:
+
+      #{showPpr term}
+
+    While trying to interpret something to type:
+
+      #{typ}
+  |]
 
 -- | Same as 'termToData', but errors hard if it can't translate a given term
 -- to data.